refactor add test form radio question
