+++ src/components/routes/my/tests/add/Form/Questions/questionTypes/Radio/Option/ActionBtns/Delete.tsx
@@ -1,0 +1,34 @@
+import { Button } from "@/components/ui/button";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
+import useAddTestFormRadioQuestionContext from "@/lib/hooks/addTestForm/radioQuestionContext";
+import { TrashIcon } from "lucide-react";
+
+type Props = {
+  optionIndex: number;
+};
+
+export default function AddTestFormRadioQuestionOptionDeleteBtn({
+  optionIndex,
+}: Props) {
+  const { optionsFields } = useAddTestFormRadioQuestionContext();
+
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild>
+        <Button
+          variant="ghost"
+          type="button"
+          className="text-muted-foreground hover:text-destructive"
+          onClick={() => optionsFields.remove(optionIndex)}
+        >
+          <TrashIcon />
+        </Button>
+      </TooltipTrigger>
+      <TooltipContent>Удалить</TooltipContent>
+    </Tooltip>
+  );
+}
