+++ src/components/routes/my/tests/add/Form/Questions/questionTypes/Radio/Option/ActionBtns/Index.tsx
@@ -1,0 +1,23 @@
+import { TooltipProvider } from "@/components/ui/tooltip";
+import { AddTestFormSavedQuestionOptionSchemaType } from "@/lib/zod/schemas/addTestForm/QuestionOption";
+import AddTestFormRadioQuestionOptionMarkAsCorrectBtn from "./MarkAsCorrect";
+import AddTestFormRadioQuestionOptionDeleteBtn from "./Delete";
+
+type Props = {
+  optionIndex: number;
+  option: AddTestFormSavedQuestionOptionSchemaType;
+};
+
+export default function AddTestFormRadioQuestionOptionActionBtns({
+  option,
+  optionIndex,
+}: Props) {
+  return (
+    <TooltipProvider>
+      <AddTestFormRadioQuestionOptionMarkAsCorrectBtn
+        {...{ option, optionIndex }}
+      />
+      <AddTestFormRadioQuestionOptionDeleteBtn {...{ optionIndex }} />
+    </TooltipProvider>
+  );
+}
