+++ src/lib/hooks/addTestForm/radioQuestionContext.ts
@@ -1,0 +1,6 @@
+import AddTestFormRadioQuestionContext from "@/lib/contexts/addTestForm/radioQuestion";
+import useContextVal from "../contextVal";
+
+export default function useAddTestFormRadioQuestionContext() {
+  return useContextVal(AddTestFormRadioQuestionContext);
+}
